src/requirements.txt
numpy==1.22.0
pandas==1.5.3
scikit-learn==1.3.0
xgboost==1.3.3
joblib==1.3.1
python-dotenv==1.0.0
azureml-core==1.49.0
azureml-dataset-runtime==1.49.0